Create README.md
This project deals with the comparison of cars. It helps the user to select the best
car of their requirement according to their criteria.
The information is fed up on the source code, allowing users to compare two cars of their criteria. This program is implemented in C++. The
source code is easy and straightforward through which the comparison between cars is easily
determined. The case study discusses the use of more straightforward codes to build abstraction.
In addition, this report will also discuss the strategies to concise the large set of data
in a few numbers of lines.
The users can interact with the program by giving input about the car of their choice and
after execution, the output gives a comparison.